feat: crea propiedad y métodos para nombre de instructor
